# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: PetStore
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /pet/winner/{pet_id}:
    x-swagger-router-controller: pet
    get:
      operationId: bidWinner
      description: display winners of bid
      # define the parameters
      parameters:
        - name: pet_id
          description: Pet id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/BidWinners"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  BidWinners:
    type: object
    properties:
      user: 
        type: string
      bid:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string